move cli commands into a tasks module
